Save on quit, display save message
When the last user leaves a document, the document is saved, unloaded from
server memory, and forgotten.

This
- Adds primitive auto-saving
- Frees memory. Otherwise the server keeps a copy of every document in memory
  indefinitely.
- Allows editing documents on disk with another program.

Additionally, on manual saves the message "Saved." is now displayed at the
bottom for clarity.

It would be good to add true intervals in the future, but I don't have the time
at the moment.